The inputs of the form should at least support the following HTML5 verification to make sure that the data entered by the users is correct.
The inputs of the form should at least support the following HTML5
verification to make sure that the data entered by the users is correct.
The "Color" field must be implemented using the input type "color".
The "Number" field must be implemented using the input type "number" and
can only contain numbers between 0 and 130.
If the data entered by the user is correct, use Web SQL technique to
save the data into a database when the users press the "SAVE" button.